+++ internal/model/balance.go
@@ -1,8 +1,8 @@
 package model
 
 type Balance struct {
-	ID           int  `json:"-"`
-	User         User `json:"-"`
-	Balance      int  `json:"current"`
-	SpentAllTime int  `json:"withdrawn"`
+	ID           int     `json:"-"`
+	User         User    `json:"-"`
+	Balance      float32 `json:"current"`
+	SpentAllTime float32 `json:"withdrawn"`
 }
